requirements.txt
attrs==21.2.0
aws_cdk.asset_awscli_v1==2.2.202
aws_cdk.asset_kubectl_v20==2.1.2
aws_cdk.asset_node_proxy_agent_v6==2.0.1
aws_cdk.assets==1.204.0
aws_cdk.aws_applicationautoscaling==1.204.0
aws_cdk.aws_autoscaling_common==1.204.0
aws_cdk.aws_cloudwatch==1.204.0
aws_cdk.aws_codeguruprofiler==1.204.0
aws_cdk.aws_codestarnotifications==1.204.0
aws_cdk.aws_ec2==1.204.0
aws_cdk.aws_ecr==1.204.0
aws_cdk.aws_ecr_assets==1.204.0
aws_cdk.aws_efs==1.204.0
aws_cdk.aws_events==1.204.0
aws_cdk.aws_iam==1.204.0
aws_cdk.aws_kms==1.204.0
aws_cdk.aws_lambda==1.204.0
aws_cdk.aws_logs==1.204.0
aws_cdk.aws_s3==1.204.0
aws_cdk.aws_s3_assets==1.204.0
aws_cdk.aws_signer==1.204.0
aws_cdk.aws_sns==1.204.0
aws_cdk.aws_sns_subscriptions==1.204.0
aws_cdk.aws_sqs==1.204.0
aws_cdk.aws_ssm==1.204.0
aws_cdk.cloud_assembly_schema==1.204.0
aws_cdk.core==1.204.0
aws_cdk.cx_api==1.204.0
aws_cdk.region_info==1.204.0
aws_cdk_lib==2.134.0
Brotli==1.0.9
ConfigParser==7.0.0
constructs==10.3.0
cryptography==3.4.8
Cython==3.0.10
dl==0.1.0
docutils==0.17.1
HTMLParser==0.0.2
importlib_metadata==4.6.4
ipython==8.12.3
ipython==7.31.1
ipywidgets==8.1.2
Jinja2==3.1.3
jnius==1.1.0
keyring==23.5.0
lockfile==0.12.2
lxml==4.8.0
numpy==1.21.5
ordereddict==1.1
Pillow==9.0.1
Pillow==10.3.0
protobuf==3.12.4
pyOpenSSL==24.1.0
railroad==0.5.0
Sphinx==7.3.5
thread==1.1.1
toml==0.10.2
tornado==6.4
xmlrpclib==1.0.1
zipp==1.0.0
